requirements.txt
asgiref==3.9.2
Django==4.2.24
django-cors-headers==4.9.0
djangorestframework==3.16
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
Unidecode==1.4.0
